use exact-width integer types

+++ src/calibration.cpp
@@ -1,6 +1,7 @@
 #include "calibration.h"
 
 #include <avr/eeprom.h>
+#include <stdint.h>
 
 #include "ui/common.h"
 
@@ -11,8 +12,8 @@ Calibration::Calibration()
 
 ui::Position Calibration::translate(ui::Position position) const
 {
-    int x = position.x;
-    int y = position.y;
+    int16_t x = position.x;
+    int16_t y = position.y;
     position.x = theta_[0][0] * x + theta_[0][1] * y + theta_[0][2];
     position.y = theta_[1][0] * x + theta_[1][1] * y + theta_[1][2];
     return position;
