+++ src/ui/display.h
@@ -1,12 +1,14 @@
 #ifndef UI_DISPLAY_H_
 #define UI_DISPLAY_H_
 
+#include <stdint.h>
+
 namespace ui {
 
 class Display {
 public:
-    virtual void data(unsigned int) const = 0;
-    virtual void address(unsigned int, unsigned int, unsigned int, unsigned int) const = 0;
+    virtual void data(uint16_t) const = 0;
+    virtual void address(uint16_t, uint16_t, uint16_t, uint16_t) const = 0;
 };
 
 } // namespace ui
